Feat: restructure the voiceagent function
