+++ platforms/macos/client/Sources/VoiceAgent/Core/JSONValue.swift
@@ -1,0 +1,48 @@
+import Foundation
+
+/// 通用 JSON 值类型，用于描述工具的参数 schema 等不规则结构。
+public indirect enum JSONValue: Sendable, Equatable {
+    case null
+    case bool(Bool)
+    case int(Int)
+    case double(Double)
+    case string(String)
+    case array([JSONValue])
+    case object([String: JSONValue])
+}
+
+extension JSONValue: Codable {
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+        if container.decodeNil() {
+            self = .null
+        } else if let b = try? container.decode(Bool.self) {
+            self = .bool(b)
+        } else if let i = try? container.decode(Int.self) {
+            self = .int(i)
+        } else if let d = try? container.decode(Double.self) {
+            self = .double(d)
+        } else if let s = try? container.decode(String.self) {
+            self = .string(s)
+        } else if let a = try? container.decode([JSONValue].self) {
+            self = .array(a)
+        } else if let o = try? container.decode([String: JSONValue].self) {
+            self = .object(o)
+        } else {
+            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Cannot decode JSONValue")
+        }
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        switch self {
+        case .null: try container.encodeNil()
+        case let .bool(b): try container.encode(b)
+        case let .int(i): try container.encode(i)
+        case let .double(d): try container.encode(d)
+        case let .string(s): try container.encode(s)
+        case let .array(a): try container.encode(a)
+        case let .object(o): try container.encode(o)
+        }
+    }
+}
